Drop Apple Pay domain registration and the payment intents concern
Domain registration presumed the backend knows the storefront's domain. Headless,
the storefront is a separate deployment on a domain Spree never sees, so
registering store.url registers the wrong host — which looks like success.
Merchants register domains in the Stripe dashboard, as a headless setup must
anyway.

A PaymentIntents concern beside PaymentSessions implied two payment paths, one
of them legacy. There is one: creating a payment session is creating a Stripe
payment intent, and the session stores the intent id as its external id. The
intent calls move into the sessions concern with payload building private to it.

PaymentIntentPresenter presented nothing — it built a request body, and
update_payment_intent constructed the whole create-payload only to discard most
of it.
