+++ docs/plans/6.0-payment-gateways-monorepo.md
@@ -4,7 +4,7 @@
 **Target:** Spree 6.0
 **Depends on:** Payment sessions infrastructure (shipped in core), `6.0-multi-vendor-marketplace.md` (Phase 3b consumes the Stripe gem)
 **Author:** Damian Legawiec
-**Last updated:** 2026-08-09
+**Last updated:** 2026-08-10
 
 ## Summary
 
@@ -136,22 +136,29 @@ session gateway, `Spree::Gateway::Bogus`, implements the same six. What left is
 the machinery *around* them: creating intents for payments that never had a
 session, off-session confirmation, and the storefront's redirect/confirm flow.
 
-**Stripe (45 files, ~1,650 LOC):**
+**Stripe (39 files, ~1,400 LOC):**
 - `SpreeStripe::Gateway` + `Gateway::PaymentSessions`,
-  `Gateway::PaymentSetupSessions`, `Gateway::PaymentIntents`,
-  `Gateway::Webhooks` concerns (intents are Stripe-side API objects wrapped by
-  the session flow — no intent AR model exists)
+  `Gateway::PaymentSetupSessions`, `Gateway::Webhooks` concerns (intents are
+  Stripe-side API objects created by the session flow — no intent AR model and
+  no separate intents concern)
 - `Spree::PaymentSessions::Stripe`, `Spree::PaymentSetupSessions::Stripe`
   (core-table STI subclasses)
 - Payment sources (Klarna, Affirm, Alipay, iDEAL, Link, SEPA, bank transfer,
   Przelewy24, AfterPay), services (CreatePayment, CreateSource,
-  CreatePaymentSession, CreateGatewayWebhooks, RegisterDomain,
-  UpdateCustomer), presenters, jobs, the `.stripe` scopes on
-  `PaymentMethod`/`GatewayCustomer`
+  CreatePaymentSession, CreateGatewayWebhooks, UpdateCustomer), jobs, the
+  `.stripe` scopes on `PaymentMethod`/`GatewayCustomer`
 
-Apple Pay / Google Pay domain **registration** with Stripe stays — quick
-checkout in the session flow depends on it. Serving the domain-association
-**file** was a Rails-storefront route and left with the storefront.
+**Apple Pay / Google Pay domain registration is dropped too** (2026-08-10).
+It assumed Spree knows the storefront's domain, which a headless install does
+not — the storefront is someone else's deployment, on a domain the backend
+never sees. Registering `store.url` there registers the wrong host. Merchants
+register their domains in the Stripe dashboard, which is where a headless setup
+has to do it anyway. This removes `RegisterDomain`, its job, the
+`CustomDomain` decorator and the domain accessors on `Store`.
+
+Stripe payment **intents** are not a separate subsystem and have no concern of
+their own: creating a payment session *is* creating an intent, so those calls
+live in `Gateway::PaymentSessions` with the payload building private to it.
 
 ### Core adaptations the port required
 
@@ -196,6 +203,10 @@ sources):**
 - Assorted AR decorators (order/payment/store/custom_domain/user) tied to the
   legacy flows
 - All `use_legacy_*` preferences and both branches of their conditionals
+- **Apple Pay / Google Pay domain registration** (2026-08-10): `RegisterDomain`,
+  `RegisterDomainJob`, the `CustomDomain` decorator and the domain accessors on
+  `Store`. It presumed the backend knows the storefront's domain, which is false
+  headless — merchants register domains in the Stripe dashboard instead.
 - **Classic-checkout payment machinery** (2026-08-09): Stripe Tax
   (`Calculators::StripeTax`, `TaxPresenter`, `CreateTaxTransactionJob`, the
   shipment decorator that fired it), `SpreeStripe::CompleteOrder` +
